feat: Refactor configuration modules to enhance structure and organization
