+++ src/lib.rs
@@ -16,36 +16,19 @@
 //! - Extensible design through the `DynamicConfigs` trait
 //! - Helper methods for common configuration tasks
 
-mod app;
-mod aws;
-mod configs;
-mod dynamo;
-mod environment;
-mod health_readiness;
-mod identity_server;
-mod influx;
-mod kafka;
-mod metrics;
-mod mqtt;
-mod postgres;
-mod rabbitmq;
-mod secrets;
-mod sqlite;
-mod traces;
-
-pub use app::AppConfigs;
-pub use aws::AwsConfigs;
-pub use configs::{Configs, DynamicConfigs, Empty};
-pub use dynamo::DynamoConfigs;
-pub use environment::Environment;
-pub use health_readiness::HealthReadinessConfigs;
-pub use identity_server::IdentityServerConfigs;
-pub use influx::InfluxConfigs;
-pub use kafka::KafkaConfigs;
-pub use metrics::{MetricConfigs, MetricExporterKind};
-pub use mqtt::{MQTTBrokerKind, MQTTConfigs, MQTTConnectionConfigs, MQTTTransport};
-pub use postgres::{PostgresConfigs, PostgresSslMode};
-pub use rabbitmq::RabbitMQConfigs;
-pub use secrets::SecretsManagerKind;
-pub use sqlite::SqliteConfigs;
-pub use traces::{TraceConfigs, TraceExporterKind};
+pub mod app;
+pub mod aws;
+pub mod configs;
+pub mod dynamic;
+pub mod dynamo;
+pub mod environment;
+pub mod health_readiness;
+pub mod identity_server;
+pub mod influx;
+pub mod kafka;
+pub mod mqtt;
+pub mod otlp;
+pub mod postgres;
+pub mod rabbitmq;
+pub mod secrets;
+pub mod sqlite;
